requirements.txt
# engine
nextcord

# other
python-dotenv

# database
sqlalchemy
